# JDK-8282625 Formatter caches Locale/DecimalFormatSymbols poorly