--- a/src/java.base/share/classes/java/util/Formatter.java
+++ b/src/java.base/share/classes/java/util/Formatter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -2008,14 +2008,27 @@ import sun.util.locale.provider.ResourceBundleBasedAdapter;
  * @since 1.5
  */
 public final class Formatter implements Closeable, Flushable {
+
+    // Caching DecimalFormatSymbols
+    static DecimalFormatSymbols DFS = null;
+    static DecimalFormatSymbols getDecimalFormatSymbols(Locale locale) {
+        DecimalFormatSymbols dfs = DFS;
+        if (dfs != null && dfs.getLocale() == locale) {
+            return dfs;
+        }
+        dfs = DecimalFormatSymbols.getInstance(locale);
+        DFS = dfs;
+        return dfs;
+    }
+
+    // Caching zero.
+    static char getZero(Locale locale) {
+        return locale == null ? '0' : getDecimalFormatSymbols(locale).getZeroDigit();
+    }
+
     private Appendable a;
     private final Locale l;
-
     private IOException lastException;
-
-    // Non-character value used to mark zero as uninitialized
-    private static final char ZERO_SENTINEL = '\uFFFE';
-    private char zero = ZERO_SENTINEL;
 
     /**
      * Returns a charset object for the given charset name.
@@ -2521,20 +2534,6 @@ public final class Formatter implements Closeable, Flushable {
      */
     public Formatter(OutputStream os, Charset charset, Locale l) {
         this(l, new BufferedWriter(new OutputStreamWriter(os, charset)));
-    }
-
-    private char zero() {
-        char zero = this.zero;
-        if (zero == ZERO_SENTINEL) {
-            if ((l != null) && !l.equals(Locale.US)) {
-                DecimalFormatSymbols dfs = DecimalFormatSymbols.getInstance(l);
-                zero = dfs.getZeroDigit();
-            } else {
-                zero = '0';
-            }
-            this.zero = zero;
-        }
-        return zero;
     }
 
     /**
@@ -4498,14 +4497,6 @@ public final class Formatter implements Closeable, Flushable {
             throw new IllegalFormatConversionException(c, arg.getClass());
         }
 
-        private char getZero(Formatter fmt, Locale l) {
-            if ((l != null) &&  !l.equals(fmt.locale())) {
-                DecimalFormatSymbols dfs = DecimalFormatSymbols.getInstance(l);
-                return dfs.getZeroDigit();
-            }
-            return fmt.zero();
-        }
-
         private StringBuilder localizedMagnitude(Formatter fmt, StringBuilder sb,
                 long value, int flags, int width, Locale l) {
             return localizedMagnitude(fmt, sb, Long.toString(value, 10), 0, flags, width, l);
@@ -4519,7 +4510,7 @@ public final class Formatter implements Closeable, Flushable {
             }
             int begin = sb.length();
 
-            char zero = getZero(fmt, l);
+            char zero = getZero(l);
 
             // determine localized grouping separator and size
             char grpSep = '\0';
@@ -4539,7 +4530,7 @@ public final class Formatter implements Closeable, Flushable {
                 if (l == null || l.equals(Locale.US)) {
                     decSep  = '.';
                 } else {
-                    DecimalFormatSymbols dfs = DecimalFormatSymbols.getInstance(l);
+                    DecimalFormatSymbols dfs = getDecimalFormatSymbols(l);
                     decSep  = dfs.getDecimalSeparator();
                 }
             }
@@ -4549,7 +4540,7 @@ public final class Formatter implements Closeable, Flushable {
                     grpSep = ',';
                     grpSize = 3;
                 } else {
-                    DecimalFormatSymbols dfs = DecimalFormatSymbols.getInstance(l);
+                    DecimalFormatSymbols dfs = getDecimalFormatSymbols(l);
                     grpSep = dfs.getGroupingSeparator();
                     DecimalFormat df = null;
                     NumberFormat nf = NumberFormat.getNumberInstance(l);
@@ -4612,7 +4603,7 @@ public final class Formatter implements Closeable, Flushable {
         // group separators is added for any locale.
         private void localizedMagnitudeExp(Formatter fmt, StringBuilder sb, char[] value,
                 final int offset, Locale l) {
-            char zero = getZero(fmt, l);
+            char zero = getZero(l);
 
             int len = value.length;
             for (int j = offset; j < len; j++) {